Moving classes to NuCache and making them internal
