+++ src/Umbraco.PublishedCache.NuCache/Navigable/INavigableFieldType.cs
@@ -1,11 +1,11 @@
-namespace Umbraco.Cms.Core.Xml.XPath;
+namespace Umbraco.Cms.Infrastructure.PublishedCache.Navigable;
 
 /// <summary>
 ///     Represents the type of a field of a content that can be navigated via XPath.
 /// </summary>
 /// <remarks>A field can be an attribute or a property.</remarks>
 [Obsolete("The current implementation of XPath is suboptimal and will be removed entirely in a future version. Scheduled for removal in v15. Still needed for NuCache")]
-public interface INavigableFieldType
+internal interface INavigableFieldType
 {
     /// <summary>
     ///     Gets the name of the field type.
