+++ src/Umbraco.PublishedCache.NuCache/Navigable/Source.cs
@@ -1,5 +1,4 @@
 using Umbraco.Cms.Core.Models.PublishedContent;
-using Umbraco.Cms.Core.Xml.XPath;
 
 namespace Umbraco.Cms.Infrastructure.PublishedCache.Navigable;
